// src/models/IPositionTranslation.ts
export interface PositionTranslation {
  no: string;
  name: string;
  description: string;
  status: string;
  action: string;
  search: string;
  clear: string;
  addProject: string;
  edit: string;
  delete: string;
  save: string;
  cancle: string;
  confirm1: string;
  confirm2: string;
  confirm3: string;
  page: string;
  status1: string;
  status2: string;
  statusFail: string;
  statusAdd: string;
  statusEdit: string;
  statusDelete: string;
  noProject: string;
  please: string;
}